# docs/api-specs.yml
openapi: 3.0.0
info:
  title: E-commerce
  description: |
    REST API для интернет-магазина с поддержкой:
    - Управления каталогом товаров и категорий
    - Системы заказов с управлением складом
    - Корзины покупок
    - Аутентификации пользователей
    - Системы отзывов и рейтингов
    
    ## Аутентификация
    API использует JWT токены. Для доступа к защищенным endpoints необходимо передавать токен в заголовке:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Пагинация
    Все endpoint'ы с множественными результатами поддерживают пагинацию:
    - `page` - номер страницы
    
    - `size` - размер страницы (по умолчанию 20)
    
  version: 0.0.1
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.ecommerce.example.com/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для аутентификации
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          description: Уникальный код ошибки
          example: "PRODUCT_NOT_FOUND"
        message:
          type: string
          description: Человекочитаемое описание ошибки
          example: "Товар с ID 123 не найден"
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: "2024-01-15T10:30:00Z"
        validationErrors:
          type: object
          additionalProperties:
            type: string
          description: Детали ошибок валидации
          example:
            email: "Некорректный формат email"
            password: "Пароль должен содержать минимум 8 символов"
    PageInfo:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          minimum: 0
          description: Текущая страница
          example: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
          description: Размер страницы
          example: 20
        totalElements:
          type: integer
          minimum: 0
          description: Общее количество элементов
          example: 150
        totalPages:
          type: integer
          minimum: 0
          description: Общее количество страниц
          example: 8
    User:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 1
        email:
          type: string
          format: email
          description: Email пользователя
          example: "john.doe@example.com"
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          description: Имя пользователя
          example: "Джон"
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          description: Фамилия пользователя
          example: "Доу"
        role:
          type: string
          enum: [USER, ADMIN]
          description: Роль пользователя
          example: "USER"
        createdAt:
          type: string
          format: date-time
          description: Дата регистрации
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления
          example: "2024-01-15T10:30:00Z"
    UsersPage:
      type: object
      required:
        - content
        - page
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/User'
        page:
          $ref: '#/components/schemas/PageInfo'
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Номер страницы (начинается с 0)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
    SizeParam:
      name: size
      in: query
      description: Размер страницы
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Получить профиль пользователя
      description: Возвращает профиль текущего авторизованного пользователя
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Users
      summary: Обновить профиль пользователя
      description: Обновляет информацию профиля текущего пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "Джон"
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                  example: "Доу"

      responses:
        '200':
          description: Профиль успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/users:
    get:
      tags:
        - Admin
      summary: Получить список всех пользователей
      description: Возвращает пагинированный список всех пользователей (только для администраторов)
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersPage'
        '403':
          description: Недостаточно прав доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...